internal/handlers/handlers_test.go
package handlers

import (
	"bytes"
	"encoding/json"
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/gin-gonic/gin"
	"github.com/tm-acme-shop/acme-shop-shared-go/models"
)

func TestHealth(t *testing.T) {
	gin.SetMode(gin.TestMode)

	h := &Handlers{}

	w := httptest.NewRecorder()
	c, _ := gin.CreateTestContext(w)

	h.Health(c)

	if w.Code != http.StatusOK {
		t.Errorf("Expected status 200, got %d", w.Code)
	}

	var resp map[string]interface{}
	if err := json.Unmarshal(w.Body.Bytes(), &resp); err != nil {
		t.Fatalf("Failed to parse response: %v", err)
	}

	if resp["status"] != "healthy" {
		t.Errorf("Expected status 'healthy', got %v", resp["status"])
	}

	if resp["service"] != "orders-service" {
		t.Errorf("Expected service 'orders-service', got %v", resp["service"])
	}
}

func TestReady(t *testing.T) {
	gin.SetMode(gin.TestMode)

	h := &Handlers{}

	w := httptest.NewRecorder()
	c, _ := gin.CreateTestContext(w)

	h.Ready(c)

	if w.Code != http.StatusOK {
		t.Errorf("Expected status 200, got %d", w.Code)
	}
}

func TestLive(t *testing.T) {
	gin.SetMode(gin.TestMode)

	h := &Handlers{}

	w := httptest.NewRecorder()
	c, _ := gin.CreateTestContext(w)

	h.Live(c)

	if w.Code != http.StatusOK {
		t.Errorf("Expected status 200, got %d", w.Code)
	}
}

func TestHandleError_NotFound(t *testing.T) {
	gin.SetMode(gin.TestMode)

	w := httptest.NewRecorder()
	c, _ := gin.CreateTestContext(w)

	handleError(c, nil) // nil error case

	// Reset for actual test
	w = httptest.NewRecorder()
	c, _ = gin.CreateTestContext(w)

	// TODO(TEAM-PLATFORM): Add actual error test cases
}

func TestCreateOrderRequestValidation(t *testing.T) {
	tests := []struct {
		name        string
		request     models.CreateOrderRequest
		shouldError bool
	}{
		{
			name: "valid request",
			request: models.CreateOrderRequest{
				UserID: "user_123",
				Items: []models.OrderItem{
					{
						ProductID:   "prod_abc",
						ProductName: "Test Product",
						Quantity:    1,
						UnitPrice:   models.Money{Amount: 1000, Currency: "USD"},
						Total:       models.Money{Amount: 1000, Currency: "USD"},
					},
				},
				ShippingAddress: models.Address{
					Line1:      "123 Test St",
					City:       "Test City",
					State:      "TS",
					PostalCode: "12345",
					Country:    "US",
				},
				BillingAddress: models.Address{
					Line1:      "123 Test St",
					City:       "Test City",
					State:      "TS",
					PostalCode: "12345",
					Country:    "US",
				},
			},
			shouldError: false,
		},
		{
			name: "missing user ID",
			request: models.CreateOrderRequest{
				Items: []models.OrderItem{
					{ProductID: "prod_abc", Quantity: 1},
				},
			},
			shouldError: true,
		},
		{
			name: "empty items",
			request: models.CreateOrderRequest{
				UserID: "user_123",
				Items:  []models.OrderItem{},
			},
			shouldError: true,
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			// TODO(TEAM-PLATFORM): Add actual validation test implementation
			_ = tt.request
			_ = tt.shouldError
		})
	}
}

func TestOrderStatusTransitions(t *testing.T) {
	tests := []struct {
		name     string
		from     models.OrderStatus
		to       models.OrderStatus
		expected bool
	}{
		{"pending to confirmed", models.OrderStatusPending, models.OrderStatusConfirmed, true},
		{"pending to cancelled", models.OrderStatusPending, models.OrderStatusCancelled, true},
		{"pending to shipped", models.OrderStatusPending, models.OrderStatusShipped, false},
		{"confirmed to processing", models.OrderStatusConfirmed, models.OrderStatusProcessing, true},
		{"processing to shipped", models.OrderStatusProcessing, models.OrderStatusShipped, true},
		{"shipped to delivered", models.OrderStatusShipped, models.OrderStatusDelivered, true},
		{"delivered to refunded", models.OrderStatusDelivered, models.OrderStatusRefunded, true},
		{"cancelled to anything", models.OrderStatusCancelled, models.OrderStatusConfirmed, false},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			// TODO(TEAM-PLATFORM): Add actual transition test implementation
			_ = tt.from
			_ = tt.to
			_ = tt.expected
		})
	}
}

func TestCreateOrderHandler(t *testing.T) {
	gin.SetMode(gin.TestMode)

	// TODO(TEAM-PLATFORM): Add integration tests with mock services
	t.Skip("Integration test - requires mock services")

	reqBody := models.CreateOrderRequest{
		UserID: "user_123",
		Items: []models.OrderItem{
			{
				ProductID:   "prod_abc",
				ProductName: "Test Product",
				Quantity:    2,
				UnitPrice:   models.Money{Amount: 1000, Currency: "USD"},
				Total:       models.Money{Amount: 2000, Currency: "USD"},
			},
		},
		ShippingAddress: models.Address{
			Line1:      "123 Test St",
			City:       "Test City",
			State:      "TS",
			PostalCode: "12345",
			Country:    "US",
		},
		BillingAddress: models.Address{
			Line1:      "123 Test St",
			City:       "Test City",
			State:      "TS",
			PostalCode: "12345",
			Country:    "US",
		},
	}

	body, _ := json.Marshal(reqBody)
	_ = bytes.NewReader(body)
}

func TestGetOrderHandler(t *testing.T) {
	gin.SetMode(gin.TestMode)

	// TODO(TEAM-PLATFORM): Add integration tests with mock services
	t.Skip("Integration test - requires mock services")
}

func TestLegacyOrderEndpoints(t *testing.T) {
	// TODO(TEAM-API): Remove these tests after v1 API migration complete
	t.Run("CreateOrderV1", func(t *testing.T) {
		t.Skip("Legacy endpoint test")
	})

	t.Run("GetOrderV1", func(t *testing.T) {
		t.Skip("Legacy endpoint test")
	})

	t.Run("ListOrdersV1", func(t *testing.T) {
		t.Skip("Legacy endpoint test")
	})
}

func TestPaymentEndpoints(t *testing.T) {
	// TODO(TEAM-PAYMENTS): Add payment endpoint tests
	t.Skip("Payment tests require mock payment client")
}

func BenchmarkCreateOrderHandler(b *testing.B) {
	gin.SetMode(gin.TestMode)

	// TODO(TEAM-PLATFORM): Add performance benchmarks
	b.Skip("Benchmark requires mock services")
}
